# DO NOT MERGE: Optimizing Hash repartition on a raw pointer, instead of using TypedValue.