--- a/catalog/Catalog.proto
+++ b/catalog/Catalog.proto
@@ -48,6 +48,14 @@ message PartitionSchemeHeader {
   extensions 16 to max;
 }
 
+message HashPartitionSchemeHeader {
+  extend PartitionSchemeHeader {
+    // Required.
+    repeated Type.TypeID partition_attr_types = 16;
+    repeated uint64 char_type_lengths = 17;
+  }
+}
+
 message PartitionValues {
   repeated TypedValue partition_values = 1;
 }