… catalog/CMakeLists.txt
… catalog/tests/NUMAPlacementScheme_unittest.cpp
… catalog/tests/PartitionScheme_unittest.cpp
… query_optimizer/resolver/CMakeLists.txt
… query_optimizer/resolver/Resolver.cpp
… query_optimizer/resolver/Resolver.hpp
… relational_operators/tests/HashJoinOperator_unittest.cpp
… storage/InsertDestination.hpp
… types/CMakeLists.txt
… types/Type.cpp
… types/TypeID.cpp
… types/TypeID.hpp