NOTE: deal with null values.

This PR optimizes hash repartition over a raw pointer, instead of `TypedValue`.

Experimental results show that repartition a `lineorder` from `SSB-100` on would reduce to `750 ms` from `1000 ms`.

Note that this PR added supports for the single partition attribute. For multi-attribute repartition, there are two basic approaches.
 1. like the single partition optimization, combine per-column hash, and then set the bit vector for the result partition.
 1. per-tuple evaluation, w/o any extra buffer used in the above approach.

The first approach on the single partition attribute actually takes `1200 ms`, while the latter takes `850 ms`.